feat: view, main 함수 작성 완료
