+++ src/main/java/ResultView.java
@@ -1,7 +1,8 @@
+import java.util.ArrayList;
 import java.util.List;
 
 public class ResultView {
-    private static void printResult(List<Car> racingCars) {
+    public static void printResult(List<Car> racingCars) {
         System.out.println("실행 결과");
         for (int i = 0; i < racingCars.size(); i++) {
             printDistance(racingCars.get(i));
@@ -20,5 +21,29 @@ public class ResultView {
             distance--;
         }
         System.out.println();
+    }
+
+    private static List<String> getWinners(List<Car> racingCars) {
+        List<String> winners = new ArrayList<>();
+
+        int maxDistance = findMaxDistance(racingCars);
+        for (Car car : racingCars) {
+            addWinner(winners, car, maxDistance);
+        }
+        return winners;
+    }
+
+    private static int findMaxDistance(List<Car> racingCars){
+        int maxDistance = -1;
+        for (Car car : racingCars) {
+            maxDistance = Math.max(maxDistance, car.getDistance());
+        }
+        return maxDistance;
+    }
+
+    private static void addWinner(List<String> winners, Car car, int maxDistance){
+        if (car.getDistance() == maxDistance) {
+            winners.add(car.getName());
+        }
     }
 }
